[REFACTOR] 1. 시그널에 관련된 함수들을 input에서 분리시킴. 2. 시그널 출력하는 부분을 새로운 함수(config)에 넣었음. 3. srcs에 들어가 있지 않은 함수들의 위치 조정
