+++ includes/signal.h
@@ -1,0 +1,22 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   signal.h                                           :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: dongyshi <dongyshi@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/04/12 20:07:09 by dongyshi          #+#    #+#             */
+/*   Updated: 2023/04/12 20:09:26 by dongyshi         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#ifndef SIGNAL_H
+# define SIGNAL_H
+
+#include <signal.h>
+
+void	sigint(void);
+void	sigquit(void);
+void	handler(int signum);
+
+#endif SIGNAL_H
